Implementation of Code Review - Nikita #2
Fisher-Yates Shuffle has been altered to add to the end of the deck instead of the beginning.

Made an option to create a new pack (resetting the pack) upon the pack being empty.

Added an extra method PrintPack() which prints the current pack at any point.

Fixed the null warnings within the Testing class.